fix(core): address PR review findings
- Move pydantic[email] to production deps (EmailStr requires email-validator)
- Remove unused pydantic-settings dependency
- Add TWD to zero-decimal currencies set
- Change discount_percent from int to float to preserve fractional Stripe values
- Narrow GDPR exception handling to only catch resource_missing errors
- Log warning when subscription has multiple line items

+++ core/stripe_saas_core/domain/subscription.py
@@ -65,7 +65,7 @@ class Subscription(BaseModel):
     plan_id: UUID
     quantity: int = 1
     promotion_code_id: str | None = None
-    discount_percent: int | None = None
+    discount_percent: float | None = None
     discount_end_at: datetime | None = None
     trial_ends_at: datetime | None = None
     current_period_start: datetime
